refactor: make task drag and drop its own utility
